Removed optional types
It turns out trying to use optional types and accessors is going
to make things complicated because of lifetimes.  It's probably
better to define "Empty" versions of objects which are set up to
minimize memory usage.   I feell like the overhead from having an
empty object is probably not worth fussing over though, so I'm
just going to leave it as is for now.

Note to self: Stop premature optimization, silly!