Changed EngineBuilder to work direclty on an Enigne
I think it makes a lot more sense for the EngineBuilder to work direclty
with the Engine instance it's building, rather than building up all the
components, then assembling the Engine from that.

Having the Engine accessible allows it to be modified before it is built,
and changes to be made in-place.