Added `timeout` to tests which run the `Engine`
If the `Engine` doesn't shut down properly, it will infinite loop.  It's not particularly difficult
to do this, and it makes debugging the problem much more difficult.  I also want to avoid the
possibility of pushing an infinite loop into CI.  To prevent issues, I'm adding a `timeout` to
ensure the test will actually fail.