Removed generic `EventQueue` from the `Engine`
There really isn't any need for users to *need* to change out the event queue
type.  In nearly all cases, you're going to want the default MPSC queue.

Having to wrangle `E: EventQueue<Event>` in *everything* that touches an
`Engine` is proving to be a massive pain in the ass.  It makes the API more
significantly harder to work with, and only has benefits in very, very specific
use-cases, so I'm removing the generic.

If you *really*, absolutely *must* change the event queue type for some reason,
you can fork the engine, and change the type yourself.