Simplified the PR template
The old one was a bit cumbersome, and annoying to complete.  A lot of the checklist, for example,
was redundant with the automated CI/CD suite.  I've removed those redundant entries, and will
instead be relying on the CI/CD system to block merges until they are completed.  It may also be
worth introducing git hooks to automatically handle tasks like running `cargo fmt`.

Because I'm also planning start using Keep a Checklist for the whole project, I've updated the
"Changes Made" section to follow the standard also.