Added code-gen module, and fixed broken `system` macro
To fix breakage when using the `system` macro, and to avoid users needing to depend on, and import
`legion` alongside Wolf Engine, I've pulled Legion's `system` macro into Wolf Engine's code-gen
crate. I've also updated the macro to use `wolf_engine::ecs` instead of trying to use `legion` in
order to make it compatible with the re-exported `legion` in the `ecs` module.

I've actually never copy-pasted code from an open source project before, so I'm not sure if this
is the correct way to credit it.  Guess I'll ask around, and If not, I'll fix it later.