Renamed `Scene` to `SceneTrait`
I don't really like this name, but I can't really think of anything better at the moment.  The
"Scene" name is going to be used for a concrete `Scene` struct, and I need to renamed one of them
to eliminate clashes.  I'll probably come back to this, and try to come up with something better.