Removed `SceneTrait` impl for `Stage`
Migrated `update()` and `render()` methods to the main `Stage` body.

I don't really see a good reason to keep the `SceneTrait` impl.  My original intention was for
`Stage` to be a special `Scene`, and to allow users to use either it, or a custom, stand-alone
`Scene`, but I don't think that really makes sense anymore.  The `Scene` API is a bit specific, and
really only makes sense in the context of being used with the `Stage`.

Either way, there's a bunch of things (return types, methods, ext.) that are useless in either
use-case.  It's confusing, and bad API design, so I'm opting to simply things, and make the `Stage`
the only option.

The `Stage` is more than capable of supporting a single-scene game, anyways, and I'm not sure
anyone would ever even use a stand-alone `Scene`.