Removed `Send` / `Sync` impl from `MainLoopResource`
The `resources` crate isn't going to work, because it *only* supports storing `Send` / `Sync`
types.

I'm going to take the idea from Legion's `Resources` type, and build something with a sync
component.  Not yet sure if I'm going to just take code from Legion or not, though, because I don't
like how it just panics when borrow rules are violated.